adde petal_FT class, derived from polyFT.
pixelized_mask method is functional, with radial profile 'arch_cos' with cosine arch in there.
Needs to add (interpolated) NW2 radial profile.

Also need to define 'Gamma' array of samples on the mask border, sampled in order to have a regular sample
on the extinction density.

Border sampling will be done with r(\theta) function, with proper sampling in \theta.